fix list without converting to np array

+++ neurobeer/tractography/distance.py
@@ -105,8 +105,8 @@ def _fiberDistance_internal(fiberMatrix1, fiberMatrix2, flip=False,
 
         label, minDist = [], []
         for i in range(fiberMatrix1.shape[1]):
-            label.append(np.argmin(distance[0]))
-            minDist.append(distance[0, label[-1].astype(int)])
+            label.append(int(np.argmin(distance[0])))
+            minDist.append(distance[0, label[-1]])
             distance = np.delete(distance, 0, axis=0)
 
         return minDist, label
